chore(dnstap source): change type of `max_frame_handling_tasks` to `usize`
Related: https://github.com/vectordotdev/vector/pull/23448#discussion_r2257245275

+++ src/sources/util/framestream.rs
@@ -8,7 +8,7 @@ use std::{
     net::SocketAddr,
     path::PathBuf,
     sync::{
-        atomic::{AtomicU32, Ordering},
+        atomic::{AtomicUsize, Ordering},
         Arc, Mutex,
     },
     time::Duration,
@@ -374,7 +374,7 @@ pub trait FrameHandler {
     fn max_frame_length(&self) -> usize;
     fn handle_event(&self, received_from: Option<Bytes>, frame: Bytes) -> Option<Event>;
     fn multithreaded(&self) -> bool;
-    fn max_frame_handling_tasks(&self) -> u32;
+    fn max_frame_handling_tasks(&self) -> usize;
     fn host_key(&self) -> &Option<OwnedValuePath>;
     fn timestamp_key(&self) -> Option<&OwnedValuePath>;
     fn source_type_key(&self) -> Option<&OwnedValuePath>;
@@ -453,7 +453,7 @@ pub fn build_framestream_tcp_source(
 
         let request_limiter = RequestLimiter::new(
             MAX_IN_FLIGHT_EVENTS_TARGET,
-            frame_handler.max_frame_handling_tasks() as usize,
+            frame_handler.max_frame_handling_tasks(),
         );
 
         listener
@@ -606,7 +606,7 @@ async fn handle_stream(
             })
         });
 
-    let active_parsing_task_nums = Arc::new(AtomicU32::new(0));
+    let active_parsing_task_nums = Arc::new(AtomicUsize::new(0));
     loop {
         let mut permit = tokio::select! {
             _ = &mut tripwire => break,
@@ -662,7 +662,7 @@ async fn handle_tcp_frame<T>(
     frame: Bytes,
     event_sink: &mut SourceSender,
     received_from: Option<Bytes>,
-    active_parsing_task_nums: Arc<AtomicU32>,
+    active_parsing_task_nums: Arc<AtomicUsize>,
 ) where
     T: TcpFrameHandler + Send + Sync + Clone + 'static,
 {
@@ -764,7 +764,7 @@ pub fn build_framestream_unix_source(
     };
 
     let fut = async move {
-        let active_parsing_task_nums = Arc::new(AtomicU32::new(0));
+        let active_parsing_task_nums = Arc::new(AtomicUsize::new(0));
 
         info!(message = "Listening...", ?path, r#type = "unix");
 
@@ -834,7 +834,7 @@ fn build_framestream_source<T: Send + 'static>(
     out: SourceSender,
     shutdown: impl Future<Output = T> + Unpin + Send + 'static,
     span: Span,
-    active_task_nums: Arc<AtomicU32>,
+    active_task_nums: Arc<AtomicUsize>,
     error_mapper: impl FnMut(std::io::Error) + Send + 'static,
 ) {
     let content_type = frame_handler.content_type();
@@ -904,8 +904,8 @@ async fn spawn_event_handling_tasks(
     event_handler: impl FrameHandler + Send + Sync + 'static,
     mut event_sink: SourceSender,
     received_from: Option<Bytes>,
-    active_task_nums: Arc<AtomicU32>,
-    max_frame_handling_tasks: u32,
+    active_task_nums: Arc<AtomicUsize>,
+    max_frame_handling_tasks: usize,
 ) -> JoinHandle<()> {
     wait_for_task_quota(&active_task_nums, max_frame_handling_tasks).await;
 
@@ -922,7 +922,7 @@ async fn spawn_event_handling_tasks(
     })
 }
 
-async fn wait_for_task_quota(active_task_nums: &Arc<AtomicU32>, max_tasks: u32) {
+async fn wait_for_task_quota(active_task_nums: &Arc<AtomicUsize>, max_tasks: usize) {
     while max_tasks > 0 && max_tasks < active_task_nums.load(Ordering::Acquire) {
         tokio::time::sleep(Duration::from_millis(3)).await;
     }
@@ -937,7 +937,7 @@ mod test {
     use std::{
         path::PathBuf,
         sync::{
-            atomic::{AtomicU32, Ordering},
+            atomic::{AtomicUsize, Ordering},
             Arc,
         },
         thread,
@@ -986,7 +986,7 @@ mod test {
         content_type: String,
         max_frame_length: usize,
         multithreaded: bool,
-        max_frame_handling_tasks: u32,
+        max_frame_handling_tasks: usize,
         extra_task_handling_routine: F,
         host_key: Option<OwnedValuePath>,
         timestamp_key: Option<OwnedValuePath>,
@@ -1101,7 +1101,7 @@ mod test {
         fn multithreaded(&self) -> bool {
             self.multithreaded
         }
-        fn max_frame_handling_tasks(&self) -> u32 {
+        fn max_frame_handling_tasks(&self) -> usize {
             self.max_frame_handling_tasks
         }
 
@@ -1135,7 +1135,7 @@ mod test {
             self.frame_handler.multithreaded()
         }
 
-        fn max_frame_handling_tasks(&self) -> u32 {
+        fn max_frame_handling_tasks(&self) -> usize {
             self.frame_handler.max_frame_handling_tasks()
         }
 
@@ -1187,7 +1187,7 @@ mod test {
             self.frame_handler.multithreaded()
         }
 
-        fn max_frame_handling_tasks(&self) -> u32 {
+        fn max_frame_handling_tasks(&self) -> usize {
             self.frame_handler.max_frame_handling_tasks()
         }
 
@@ -1747,9 +1747,9 @@ mod test {
         let (out, rx) = SourceSender::new_test();
 
         let max_frame_handling_tasks = 20;
-        let active_task_nums = Arc::new(AtomicU32::new(0));
+        let active_task_nums = Arc::new(AtomicUsize::new(0));
         let active_task_nums_copy = Arc::clone(&active_task_nums);
-        let max_task_nums_reached = Arc::new(AtomicU32::new(0));
+        let max_task_nums_reached = Arc::new(AtomicUsize::new(0));
         let max_task_nums_reached_copy = Arc::clone(&max_task_nums_reached);
 
         let mut join_handles = vec![];
